+++ app/mailers/map_mailer.rb
@@ -1,10 +1,7 @@
 # frozen_string_literal: true
 class MapMailer < ApplicationMailer
+  include MapMailerHelper
   default from: 'team@metamaps.cc'
-
-  def access_approved_subject(map)
-    map.name + ' - access approved'
-  end
 
   def access_approved(request)
     @request = request
@@ -12,18 +9,10 @@ class MapMailer < ApplicationMailer
     mail(to: request.user.email, subject: access_approved_subject(@map))
   end
 
-  def access_request_subject(map)
-    map.name + ' - request to edit'
-  end
-
   def access_request(request)
     @request = request
     @map = request.map
     mail(to: @map.user.email, subject: access_request_subject(@map))
-  end
-
-  def invite_to_edit_subject(map)
-    map.name + ' - invited to edit'
   end
 
   def invite_to_edit(user_map)
