+++ app/mailers/topic_mailer.rb
@@ -1,19 +1,12 @@
 # frozen_string_literal: true
 class TopicMailer < ApplicationMailer
+  include TopicMailerHelper
   default from: 'team@metamaps.cc'
-
-  def added_to_map_subject(topic, map)
-    topic.name + ' was added to map ' + map.name
-  end
 
   def added_to_map(event, user)
     @entity = event.eventable
     @event = event
     mail(to: user.email, subject: added_to_map_subject(@entity, event.map))
-  end
-
-  def connected_subject(topic)
-    'new synapse to topic ' + topic.name
   end
 
   def connected(synapse, topic, user)
