+++ app/views/topic_mailer/connected.html.erb
@@ -8,7 +8,7 @@
   connected topic <span style="font-weight: bold"><%= topic1.name %></span>
   to topic <span style="font-weight: bold"><%= topic2.name %></span>
   <% if synapse.desc.length > 0 %>
-    with the description "<%= event.desc %>".
+    with the description "<%= synapse.desc %>".
   <% end %>
 </p>
